+++ lib/logger.js
@@ -1,3 +1,4 @@
+'use strict';
 const FileStreamRotator = require('file-stream-rotator');
 const logger = require('winston');
 const yaml = require('js-yaml');
