+++ lib/smtpTransport.js
@@ -1,3 +1,4 @@
+'use strict';
 const nodemailer = require('nodemailer');
 
 const fs = require('fs');
