+++ lib/utils.js
@@ -1,3 +1,4 @@
+'use strict';
 const fs = require('fs');
 
 function readFileAsString (filename, callback) {
